refactor: move all policy code to bitrouter-core
Move PolicyCache, GuardedToolRegistry, and load_policies() from
bitrouter-accounts to bitrouter-core::policy. load_policies() now
returns skipped files as data (SkippedPolicy) instead of logging
via tracing, following the core convention of no tracing dep.

Delete bitrouter-accounts::policy entirely — it was all re-exports.
