--- bitrouter-accounts/src/policy/config.rs
@@ -1,3 +1,0 @@
-//! Re-exports policy configuration types from [`bitrouter_core::policy`].
-
-pub use bitrouter_core::policy::ToolProviderPolicy;
